Nette 3.1 upgrade

+++ app/components/Media/ImageBrowserControl.php
@@ -6,18 +6,15 @@ use App\Model\Category;
 use App\Model\IO;
 use Caloriscz\Utilities\PagingControl;
 use Nette\Application\UI\Control;
-use Nette\Database\Context;
+use Nette\Database\Explorer;
 use Nette\Utils\Paginator;
 
 class ImageBrowserControl extends Control
 {
 
-    /**
-     * @var Context
-     */
-    public $database;
+    public Explorer $database;
 
-    public function __construct(Context $database)
+    public function __construct(Explorer $database)
     {
         $this->database = $database;
 
@@ -38,10 +35,13 @@ class ImageBrowserControl extends Control
     {
         $imageDb = $this->database->table('pictures')->get($id);
 
-        IO::remove(APP_DIR . '/pictures/' . $imageDb->pages_id . '/' . $imageDb->name);
-        IO::remove(APP_DIR . '/pictures/' . $imageDb->pages_id . '/tn/' . $imageDb->name);
+        if ($imageDb === null) {
+            IO::remove(APP_DIR . '/pictures/' . $imageDb->pages_id . '/' . $imageDb->name);
+            IO::remove(APP_DIR . '/pictures/' . $imageDb->pages_id . '/tn/' . $imageDb->name);
 
-        $imageDb->delete();
+            $imageDb->delete();
+        }
+
 
         $this->redirect('this', [
             'id' => $imageDb->pages_id,
